// apps/frontend/src/react/api/generated/models/ExamplePresetDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ExamplePresetDto = {
    /**
     * 프리셋 ID
     */
    id: number;
    /**
     * 프리셋 이름
     */
    name: string;
    /**
     * 프리셋 설명
     */
    description: Record<string, any> | null;
    /**
     * 프리셋에 포함된 언어 목록
     */
    languages: Array<string>;
};