// apps/frontend/src/react/api/generated/models/CreateExamplePresetRequestDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateExamplePresetRequestDto = {
    /**
     * 새 프리셋 이름
     */
    name: string;
    /**
     * 프리셋 설명
     */
    description: Record<string, any> | null;
    /**
     * 언어별 번역 예제 목록
     */
    examples?: Record<string, any>;
};